Move `Value`, `Address`, and `ElementPath` to ast/.
These are used by the AST in lots of ways, and this resolves various
layering issues.

This means that `AllocationId` also lives in ast/, but is managed by
interpreter/. A better layering here would be desirable, but this seems
good enough for the time being.
