+++ explorer/ast/BUILD
@@ -43,18 +43,26 @@ cc_library(
         "expression.cpp",
         "pattern.cpp",
         "statement.cpp",
+        "value.cpp",
     ],
     hdrs = [
+        "address.h",
         "ast.h",
         "bindings.h",
         "declaration.h",
         "element.h",
+        "element_path.h",
         "expression.h",
         "impl_binding.h",
         "pattern.h",
         "return_term.h",
         "statement.h",
+        "value.h",
         "value_node.h",
+        "value_transform.h",
+    ],
+    textual_hdrs = [
+        "value_kinds.def",
     ],
     deps = [
         ":ast_node",
@@ -136,7 +144,6 @@ cc_test(
         ":paren_contents",
         "//common:gtest_main",
         "//explorer/common:arena",
-        "//explorer/interpreter:action_and_value",
         "@com_google_googletest//:gtest",
         "@llvm-project//llvm:Support",
     ],
