+++ explorer/interpreter/builtins.h
@@ -12,9 +12,9 @@
 #include "common/error.h"
 #include "explorer/ast/declaration.h"
 #include "explorer/ast/expression.h"
+#include "explorer/ast/value.h"
 #include "explorer/common/nonnull.h"
 #include "explorer/common/source_location.h"
-#include "explorer/interpreter/value.h"
 
 namespace Carbon {
