+++ explorer/interpreter/interpreter.h
@@ -14,10 +14,10 @@
 #include "explorer/ast/declaration.h"
 #include "explorer/ast/expression.h"
 #include "explorer/ast/pattern.h"
+#include "explorer/ast/value.h"
 #include "explorer/interpreter/action.h"
 #include "explorer/interpreter/heap.h"
 #include "explorer/interpreter/trace_stream.h"
-#include "explorer/interpreter/value.h"
 #include "llvm/ADT/ArrayRef.h"
 
 namespace Carbon {
